requirements.txt
discord.py
openai
requests
zenpy
weaviate-client
tiktoken 
langchain
trafilatura